demos/sparse-mpt-7b-gsm8k/requirements.txt
deepsparse-nightly==1.6.0.20231007
transformers
gradio